chore(tests): re-unify events
