+++ tests/events.test.tsx
@@ -1,0 +1,187 @@
+import * as React from 'react'
+import { vi, it, expect, beforeEach } from 'vitest'
+import { render, fireEvent, cleanup, act as domAct } from '@testing-library/react'
+import { reconciler, Canvas } from 'react-ogl'
+
+const act = async (cb: () => Promise<void>) => domAct(async () => reconciler.act(cb))
+
+beforeEach(() => {
+  cleanup()
+})
+
+it('handles all interactive meshes', async () => {
+  const canvas = React.createRef<HTMLCanvasElement>()
+  const handleOnClick = vi.fn()
+
+  await act(async () => {
+    render(
+      <Canvas ref={canvas}>
+        <transform>
+          <mesh scale={2} onClick={handleOnClick}>
+            <box />
+            <normalProgram />
+          </mesh>
+        </transform>
+      </Canvas>,
+    )
+  })
+
+  const event = new MouseEvent('click')
+  ;(event as any).offsetX = 640
+  ;(event as any).offsetY = 400
+
+  fireEvent(canvas.current, event)
+
+  expect(handleOnClick).toHaveBeenCalled()
+})
+
+it('handles onClick', async () => {
+  const canvas = React.createRef<HTMLCanvasElement>()
+  const handleOnClick = vi.fn()
+
+  await act(async () => {
+    render(
+      <Canvas ref={canvas}>
+        <mesh scale={2} onClick={handleOnClick}>
+          <box />
+          <normalProgram />
+        </mesh>
+      </Canvas>,
+    )
+  })
+
+  const event = new MouseEvent('click')
+  ;(event as any).offsetX = 640
+  ;(event as any).offsetY = 400
+
+  fireEvent(canvas.current, event)
+
+  expect(handleOnClick).toHaveBeenCalled()
+})
+
+it('handles onPointerUp', async () => {
+  const canvas = React.createRef<HTMLCanvasElement>()
+  const handlePointerUp = vi.fn()
+
+  await act(async () => {
+    render(
+      <Canvas ref={canvas}>
+        <mesh scale={2} onPointerUp={handlePointerUp}>
+          <box />
+          <normalProgram />
+        </mesh>
+      </Canvas>,
+    )
+  })
+
+  const event = new PointerEvent('pointerup')
+  ;(event as any).offsetX = 640
+  ;(event as any).offsetY = 400
+
+  fireEvent(canvas.current, event)
+
+  expect(handlePointerUp).toHaveBeenCalled()
+})
+
+it('handles onPointerDown', async () => {
+  const canvas = React.createRef<HTMLCanvasElement>()
+  const handlePointerDown = vi.fn()
+
+  await act(async () => {
+    render(
+      <Canvas ref={canvas}>
+        <mesh scale={2} onPointerDown={handlePointerDown}>
+          <box />
+          <normalProgram />
+        </mesh>
+      </Canvas>,
+    )
+  })
+
+  const event = new PointerEvent('pointerdown')
+  ;(event as any).offsetX = 640
+  ;(event as any).offsetY = 400
+
+  fireEvent(canvas.current, event)
+
+  expect(handlePointerDown).toHaveBeenCalled()
+})
+
+it('handles onPointerMove', async () => {
+  const canvas = React.createRef<HTMLCanvasElement>()
+  const handlePointerMove = vi.fn()
+
+  await act(async () => {
+    render(
+      <Canvas ref={canvas}>
+        <mesh scale={2} onPointerMove={handlePointerMove}>
+          <box />
+          <normalProgram />
+        </mesh>
+      </Canvas>,
+    )
+  })
+
+  const event = new PointerEvent('pointermove')
+  ;(event as any).offsetX = 640
+  ;(event as any).offsetY = 400
+
+  fireEvent(canvas.current, event)
+
+  expect(handlePointerMove).toHaveBeenCalled()
+})
+
+it('handles onPointerOver', async () => {
+  const canvas = React.createRef<HTMLCanvasElement>()
+  const handleOnPointerOver = vi.fn()
+
+  await act(async () => {
+    render(
+      <Canvas ref={canvas}>
+        <mesh scale={2} onPointerOver={handleOnPointerOver}>
+          <box />
+          <normalProgram />
+        </mesh>
+      </Canvas>,
+    )
+  })
+
+  const event = new PointerEvent('pointermove')
+  ;(event as any).offsetX = 640
+  ;(event as any).offsetY = 400
+
+  fireEvent(canvas.current, event)
+
+  expect(handleOnPointerOver).toHaveBeenCalled()
+})
+
+it('handles onPointerOut', async () => {
+  const canvas = React.createRef<HTMLCanvasElement>()
+  const handlePointerOut = vi.fn()
+
+  await act(async () => {
+    render(
+      <Canvas ref={canvas}>
+        <mesh scale={2} onPointerOut={handlePointerOut}>
+          <box />
+          <normalProgram />
+        </mesh>
+      </Canvas>,
+    )
+  })
+
+  // Move pointer over mesh
+  const event = new PointerEvent('pointermove')
+  ;(event as any).offsetX = 640
+  ;(event as any).offsetY = 400
+  fireEvent(canvas.current, event)
+
+  // Move pointer away from mesh
+  const event2 = new PointerEvent('pointermove')
+  ;(event2 as any).offsetX = 0
+  ;(event2 as any).offsetY = 0
+
+  fireEvent(canvas.current, event2)
+
+  expect(handlePointerOut).toHaveBeenCalled()
+})
